\name{bcrm-package}
\alias{bcrm-package}
\docType{package}
\title{
Bayesian continual reassessment method (CRM) designs for Phase I dose escalation trials
}
\description{
Implements a wide variety of Bayesian CRM designs. The program can run interactively, allowing the user to enter outcomes after each cohort has been recruited, or via simulation to assess operating characteristics.
}
\details{
\tabular{ll}{
Package: \tab bcrm\cr
Type: \tab Package\cr
Version: \tab 0.4.3\cr
Date: \tab 2013-09-10\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}


}
\author{
Michael Sweeting \email{michael.sweeting@mrc-bsu.cam.ac.uk}

Graham Wheeler \email{graham.wheeler@mrc-bsu.cam.ac.uk}

Maintainer: Michael Sweeting \email{michael.sweeting@mrc-bsu.cam.ac.uk}
}

\references{
 Sweeting M., Mander A., Sabin T. \pkg{bcrm}: Bayesian Continual Reassessment Method Designs for Phase I Dose-Finding Trials. \emph{Journal of Statistical Software} (2013) 54: 1--26. 
 \url{http://www.jstatsoft.org/v54/i13/}
}