\name{bcrm-package}
\alias{bcrm-package}
\docType{package}
\title{
Bayesian continuous reassessment method (CRM) designs for Phase I dose escalation trials
}
\description{
Implements a wide variety of Bayesian CRM designs. The program can run interactively, allowing the user to enter outcomes after each cohort has been recruited, or via simulation to assess operating characteristics.
}
\details{
\tabular{ll}{
Package: \tab bcrm\cr
Type: \tab Package\cr
Version: \tab 0.4.1\cr
Date: \tab 2012-10-24\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}


}
\author{
Michael Sweeting \email{michael.sweeting@mrc-bsu.cam.ac.uk}

Graham Wheeler \email{graham.wheeler@mrc-bsu.cam.ac.uk}

Maintainer: Michael Sweeting \email{michael.sweeting@mrc-bsu.cam.ac.uk}
}
