% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcrm.R
\name{plot.bcrm.sim}
\alias{plot.bcrm.sim}
\title{Plot the operating characteristics from the simulated trials}
\usage{
\method{plot}{bcrm.sim}(x, trajectories = FALSE, file = NULL,
  threep3 = FALSE, ...)
}
\arguments{
\item{x}{An object of class "bcrm.sim",  as returned by \code{\link{bcrm}}
when conducting a simulation.}

\item{trajectories}{Should a summary plot of the trajectories of
administered dose levels be plotted? Defaults to FALSE.}

\item{file}{File name where the operating characteristic plot is stored,  in
a pdf format.}

\item{threep3}{Should operating characteristics of a standard 3+3 rule-based
design be plotted alongside the \code{bcrm} design? Defaults to
\code{FALSE}.}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Plots of the operating characteristics obtained from a CRM simulation.
}
\details{
This function plots the sample size distribution (if variable),  the
experimentation distribution,  the recommended dose distribution and the
percentage of subjects who experience the toxicity outcome (dose-limiting
toxicity). If \code{trajectories = TRUE} then summary statistics of
administered dose levels for each patient are plotted instead. If
\code{threep3 = TRUE} then the operating characteristics of the standard 3+3
design are plotted alongside those of the \code{bcrm} design (see
\code{\link{threep3}} for more details).
}
\references{
Sweeting M.,  Mander A.,  Sabin T. \pkg{bcrm}: Bayesian Continual
Reassessment Method Designs for Phase I Dose-Finding Trials. \emph{Journal
of Statistical Software} (2013) 54: 1--26.
\url{http://www.jstatsoft.org/article/view/v054i13}
}
\seealso{
\code{\link{print.bcrm.sim}},  \code{\link{bcrm}}, 
\code{\link{threep3}}
}
\author{
Michael Sweeting \email{mjs212@medschl.cam.ac.uk} (University of
Cambridge,  UK)
}
